+++ Jidlo.hpp
@@ -15,7 +15,7 @@ class Jidlo : public Pokrm
     string seznamIngr;
 
 public:
-    Jidlo(int id, string nazev, int cena, int trida, string ingredience);
+    Jidlo(int id, string nazev, double cena, int trida, string ingredience);
     ~Jidlo();
     void ulozIngredience(string ingredience);
     string vypisIngredience();
